+++ .idea/fileTemplates/internal/module-info.java
@@ -1,0 +1,4 @@
+#parse("License Header.java")
+#parse("File Header.java")
+module #[[$MODULE_NAME$]]# {
+}